requirements.txt
Pillow
piexif
mutagen
ffmpeg-python
tqdm
